pyproject.toml
[project]
name = "dissertation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-cloud-storage>=3.1.0",
    "httpx>=0.28.1",
    "tqdm>=4.67.1",
]
